% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_packages.R
\name{cond_packages}
\alias{cond_packages}
\title{Conditional Package Installer}
\usage{
cond_packages(packages)
}
\arguments{
\item{packages}{A list of packages needed for script that are checked to see if on machine}
}
\description{
This function conditionally installs packages when not installed on a compter.
}
\examples{
packages <- c("rgdal","mapview")
cond_packages(packages)
}
\keyword{packages}
